requirements.txt
pillow~=11.2.1
opencv-python~=4.11.0.86
numpy~=2.2.3
pypdfium2~=4.30.1
pillow-avif-plugin~=1.5.1
requests~=2.32.3
selenium~=4.33.0
bs4~=0.0.2
beautifulsoup4~=4.13.3